# public/swagger/swagger.json
openapi: 3.0.0
info:
  title: 'API Documentation'
  description: 'Dokumentasi API untuk aplikasi Akunurse'
  version: 1.0.0
paths:
...
  /upload-ijazah:
    post:
      tags:
        - 'Upload Ijazah'
      summary: 'Unggah file (PDF, JPG, JPEG, PNG) untuk Asesi'
      description: 'Mengunggah file dan menyimpannya ke dalam sistem. File yang diunggah harus memiliki format tertentu dan ukuran maksimum.'
      operationId: UploadFileIjazah
      requestBody:
        description: 'Kirim file atau Upload File Ijazah.'
        required: true
        content:
          application/json:
            schema:
              required:
                - path_file
              properties:
                path_file:
                  type: string
                  format: binary
              type: object
      responses:
        '200':
          description: 'File berhasil diunggah dan disimpan.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: 'File uploaded successfully.' }
                  data: { properties: { ijazah_id: { type: integer, example: '19' }, user_id: { type: integer, example: '10' }, path_file: { type: string, example: 'http://app.rsimmanuel.net:9091/storage/ijazah/1615203532_filename.pdf' }, valid: { type: boolean, example: null }, authentic: { type: boolean, example: null }, current: { type: boolean, example: null }, sufficient: { type: boolean, example: null }, ket: { type: string, example: 'null' } }, type: object }
                type: object
        '400':
          description: 'Validasi gagal atau file tidak diunggah.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Validasi gagal. Pastikan file di Upload.' }
                  errors: { properties: { path_file: { type: array, items: { type: string, example: 'File is required.' } } }, type: object }
                  status_code: { type: integer, example: 400 }
                type: object
        '500':
          description: 'Terjadi kesalahan tak terduga di server.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Terjadi kesalahan tak terduga di server.' }
                  error: { type: string, example: 'Error message' }
                type: object
  '/update-ijazah/{nik}':
    put:
      tags:
        - Asesor
      summary: 'Perbarui Penilaian file Ijazah berdasarkan nik untuk Asesor Terhadap Asesi'
      description: 'Memperbarui informasi detail file berdasarkan nik yang diberikan. Hanya parameter opsional yang akan diperbarui (valid, authentic, current, sufficient, ket).'
      operationId: updateFile
      parameters:
        -
          name: id
          in: path
          description: 'nik file yang ingin diperbarui'
          required: true
          schema:
            type: integer
            example: '12345678'
      requestBody:
        description: 'Kirim parameter yang ingin diperbarui. Parameter ini opsional, hanya kirim yang ingin diubah.'
        required: true
        content:
          application/json:
            schema:
              properties:
                valid:
                  type: boolean
                  example: true
                authentic:
                  type: boolean
                  example: true
                current:
                  type: boolean
                  example: true
                sufficient:
                  type: boolean
                  example: true
                ket:
                  type: string
                  example: 'Perbarui keterangan tambahan.'
              type: object
      responses:
        '200':
          description: 'File details berhasil diperbarui.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: 'File details updated successfully.' }
                  data: { properties: { path_file: { type: string, example: 'http://app.rsimmanuel.net:9091/storage/ijazah/1615203532_filename.pdf' }, valid: { type: boolean, example: true }, authentic: { type: boolean, example: true }, current: { type: boolean, example: true }, sufficient: { type: boolean, example: true }, ket: { type: string, example: 'Perbarui keterangan tambahan.' } }, type: object }
                type: object
        '400':
          description: 'Validasi gagal atau input tidak lengkap.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Validasi gagal. Pastikan semua data sudah sesuai.' }
                  errors: { properties: { valid: { type: array, items: { type: string, example: 'The valid field is required.' } } }, type: object }
                  status_code: { type: integer, example: 400 }
                type: object
        '404':
          description: 'File dengan ID yang diberikan tidak ditemukan.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'File tidak ditemukan. Silakan periksa ID dan coba lagi.' }
                  status_code: { type: integer, example: 404 }
                type: object
        '500':
          description: 'Terjadi kesalahan tak terduga saat memperbarui detail file.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Terjadi kesalahan tak terduga saat memperbarui detail file.' }
                  error: { type: string, example: 'Error message' }
                type: object
  '/ijazah/file/{nik}':
    delete:
      tags:
        - 'Upload Ijazah'
      summary: 'Menghapus file ijazah berdasarkan NIK'
      description: 'Menghapus file ijazah dari sistem berdasarkan NIK yang telah didaftarkan. Jika file ditemukan, maka akan dihapus baik dari storage dan database.'
      operationId: deleteFile
      parameters:
        -
          name: id
          in: path
          description: 'NIK file ijazah yang ingin dihapus'
          required: true
          schema:
            type: integer
            example: '12345678'
      responses:
        '200':
          description: 'File berhasil dihapus.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  message: { type: string, example: 'File berhasil dihapus dari sistem.' }
                  details: { properties: { ijazah_id: { type: integer, example: 19 }, file_path: { type: string, example: ijazah/1615203532_filename.pdf }, storage_status: { type: string, example: 'File dihapus dari storage.' }, database_status: { type: string, example: 'Rekaman file telah dihapus dari database.' } }, type: object }
                  status_code: { type: integer, example: 200 }
                type: object
        '400':
          description: 'ID yang diberikan tidak valid.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'ID tidak valid. Mohon masukkan ID berupa angka positif.' }
                  details: { properties: { id: { type: integer, example: 0 }, reason: { type: string, example: 'ID tidak memenuhi kriteria yang diperlukan.' } }, type: object }
                  status_code: { type: integer, example: 400 }
                type: object
        '404':
          description: 'File tidak ditemukan berdasarkan ID.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'File tidak ditemukan. Mohon periksa ID dan coba lagi.' }
                  details: { properties: { id: { type: integer, example: 123 }, reason: { type: string, example: 'File dengan ID ini tidak ada dalam sistem.' } }, type: object }
                  status_code: { type: integer, example: 404 }
                type: object
        '500':
          description: 'Terjadi kesalahan pada server saat menghapus file.'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  message: { type: string, example: 'Terjadi kesalahan tak terduga saat mencoba menghapus file.' }
                  error_details: { properties: { error_message: { type: string, example: 'Error message from server' }, suggestion: { type: string, example: 'Mohon periksa log server untuk informasi lebih lanjut.' } }, type: object }
                  status_code: { type: integer, example: 500 }
                type: object
  '/storage/ijazah/{path}':
    get:
      tags:
        - 'Upload Ijazah'
      summary: 'Memperlihatkan file ijazah'
      description: 'Endpoint ini digunakan untuk memperlihatkan file ijazah berdasarkan path yang diberikan.'
      operationId: getIjazah
      parameters:
        -
          name: path
          in: path
          description: 'Path file ijazah yang ingin diperlihatkan.'
          required: true
          schema:
            type: string
            example: 1615203532_filename.pdf
      responses:
        '200':
          description: 'Gambar Atau File Akan Ditampilkan.'
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: 'File tidak ditemukan.'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'File tidak ditemukan.' }
                type: object
        '500':
          description: 'Terjadi kesalahan server.'
          content:
            application/json:
              schema:
                properties:
                  error: { type: string, example: 'Terjadi kesalahan pada server.' }
                type: object
...